Fix #4298:  Limit the number of terms in the atan sum
In fpatan, for the case |x| <= 1/2, we use a Taylor series to compute
atan.  The criteria for termination is if the previous sum equals the
current sum.  This happens when the additional term is so small that
it doesn't contribute anymore.  But I think round-off causes the
previous sum and current sum to alternate so they're never equal.

To fix this, compute a rough estimate of the number of terms needed so
that the relative error between the term and x is less than a
tolerance (2^(-fpprec)).  With this change, the series then terminates
because we reach the limit of the number of terms.

For the test case, we want the atan of -7.7037b-34.  The last term
computed is 9.888b-1756, which is way smaller than needed, but at
least we terminate the loop instead of looping forever.