Fix 4299: complex-atanh error
The funny error "HOW DID I GET HERE" is a logic mistake where we
didn't handle the case where y = 0 and |x| <= 1.  From the formula in
the comment we have imagpart(atanh(z))=1/2*(arg(1+x) - arg(1-x)).  The
original code only handled |x| > 1, but clearly the formula is 0 if
|x| <= 1.

Replace the call to `error` with a bigfloat 0.

Manually tested the error case and we don't get an error.  The answer
matches what `atanh(- 3.378259464834472b-1)` returns in the repl.