Don't set special vars dn* and nn* in numden-var
Simplify usage of numden-var by changing it so that it no longer sets
the special variables nn* and dn*.  These values are also returned as
the values of numden-var (but weren't actually because of a bug).

Hence, modify all callers to accept the values of numden-var and set
the special variables there, if needed.  Currently, we always do
because we didn't investigate if this was really necessary.

Full testsuite + share passes with cmucl and ecl.