Reduce use of special vars nn* and dn*
In `kindp34`, we don't need to bind the special vars `nn*` and `dn*`.
Since `dn*` is immediately assigned to `d`, we can just do that
without binding `dn*` first.  Also, it appears the `nn*` is not used,
so we don't even need to bind `nn*` at all.  However, `kindp34` isn't
called at all in the testsuite so it's hard to know.

In `ssp`, we can do the same because `nn*` is immediately assigned to
`u`.  And AFAICT, we don't need to bind `dn*` at all.  So rename `dn*`
to `dn` in `ssp`.

Full testsuite + share tests passes with cmucl, ecl, and clisp.