zn does not need to be a special variable
The variable `zn` is set in `ggr` but is never used there.  It is set
and used in `maybpc` and only set and read in a small part of
`maybpc`.

Thus, remove the declaration of `zn` as a special variable, remove it
from `ggr` and make it lexical in `maybpc`.

Full testsuite + share passes with cmucl and ccl64.