Get rid of special var nd*
The special variable `nd*` is accessed in `maybpc` in exactly one
place.  And `maybpc` is only called from `ggr1` which also happens to
set `nd*`.  Finally, `ggr` initializes `nd*` which is the only caller
of `ggr1`.  We can test the use of `nd*` in `maybpc` using
`integrate(sin(s*x)/sqrt(x),x,0,inf)`, which is a test in
rtestint.mac, line 545.

Based on this, we can get rid of the special variable `nd*` by having
`ggr1` and `maybpc` take an extra arg for the value of `nd*`, and update
callers appropriately to pass the value of `nd*`.

Full testsuite + share passes with cmucl and ccl64.