Remove special variable exp1
The special variable `expq` was set in `intcv3`.  Change intcv3 to use
a lexical and return the value that was previously assigned to `exp1`.

`intcv3` is called only from `intcv2`, so modify `intcv2` to accept
the return value from `intcv3`.  Also move `intcv3` into `intcv2`
since it's not called from anywhere else.  We just left it as an
`flet` function even though it could very easily be inlined.  This is
just to make it match the old code in some sense.

`intcv2` takes the value from `intcv3` and can modify it.  Thus have
`intcv2` return this modified value.  It appears that the only caller
of `intcv2` that uses the return value is `intcv1`, so update `intcv1`
to save the value so it can be passed to `defint`.

Full testsuite + share passes with cmucl and ccl64 (ignoring some
newly introduced errors about `thru` and `signum`.