Remove special variables *ul1* and *ll1*
These special variables are used as the new upper and lower limits for
integration when a change of variables to `'yx` is done in `intcv2`.

`intcv2` computes the new upper and lower limits which are used in
`intcv1`.  To remove the special variables, have `intcv2` return the
new limits as additional values.  `intcv1` accepts these extra values,
validates them as before and the calls `defint` with the new integrand
and the new limits.

Full testsuite + share passes with cmucl and ccl64.