Rename *ll* and *ul* to ll and ul in interval-list
The function `interval-list` doesn't really need the special variables
so rename them so as not to use the special variables.

However, `interval-list` previously modified the values of `*ll*` and
`*ul*` so in renaming, we need the function to return the modified
values as extra return values.  The only caller (`take-principal`) is
updated to accept these extra values and updates `*ll*` and `*ul*` there.

Full testsuite + share passes with cmucl and ccl64.