Rename *ll* and *ul* to ll and ul in take-principal
The function `take-principal` doesn't really need the special
variables so rename them so as not to use the special variables.

(This also suggests that the modified values of `*ll*` and `*ul*` from
`interval-list` are never used again.  Needs to be verified, though;
this change doesn't investigate that idea.)

Full testsuite + share passes with cmucl and ccl64.