transl: fix toplevel eval_when for eval and load times
In Macsyma there was a special case for inserting an EVAL-WHEN into
translated code for COMPILE time.  In Maxima 5.0 a new case was
present for handling COMPILE, EVAL and LOAD times, while keeping
the original COMPILE time case as well (even though it then became
unused).  The names COMPILE, EVAL and LOAD were old Maclisp names
and later became deprecated in CL.

The EVAL and LOAD time cases became broken by commit 26594498 in
2007 which seemed to blindly change the deprecated names to the
non-deprecated names without understanding what the code was doing.
The COMPILE time case only continued to work because it fell back
to the original code for the COMPILE time case.

Now the translator again handles a toplevel eval_when with compile,
eval and load times, and in the translations we now use the
non-deprecated names :COMPILE-TOPLEVEL, :EXECUTE and :LOAD-TOPLEVEL.

The case of eval_when with translate time has continued to work, as
it was handled separately.

No problems with the test suite, share test suite or rtest_translator.