Add ll and ul args to order-limits
`order-limits` previously accessed `*ll*` and `*ul*` and could swap
the values.  To remove the dependency on these special variables, add
`ll` and `ul` arguments.  However, since `order-limits` can change the
order, we return the values of `ll` and `ul` as additional arguments.

`make-defint-assumptions` calls `order-limits`, so we need to pass in
the additional values and accept the additional values, modifying
`*ll*` and `*ul*` appropriately for now.

`principal-value-integral` also calls `order-limits`.  We have to
handle the possible reordering here so that we can call
`take-principal` with the correct limits of integration.