Update rtestint problem 237 to match the actual result.
The actual result differs slightly from the expected result, but
they're equivalent.  Curiously, the comment in the problem gives our
actual result, so update the expected expected result with our actual
result that matches the comment.  Remove the comment too.

With this change, the testsuite + share tests pass with cmucl and
ccl64.

This change finalizes the removal of `*ll*` and `*ul*` from
defint.lisp.