Rename specvar integer-info to *integer-info*
First, rename `integer-info` to `*integer-info*` to make it clear that
it's a special variable.

Second, since `*integer-info*` is used in several files, add a
`defvar` for it in globals.lisp and remove `integer-info` from any
special declarations.

Finally, limit.lisp also has the special variable `old-integer-info`.
Rename to `*old-integer-info*` and add a `defvar` for it, removing the
`declare-top` declaration.

Full testsuite + share passes with cmucl and ccl64.