Simpilify how print-help-string works and support gcl
Instead of the hand-written code, use a format string (modified from
the CLHS) to wrap the help string neatly.

And gcl 2.6.14 works with this code.  However, in case 2.6.13 doesn't,
we've added a defvar to control whether wrapping the help string is
done or not.  It defaults to wrapping, of course.