Add support for external html docs
WIP.

First, use keyword args for `$build_and_dump_html_index`.  Then add
new keywords:

* output_file: specifies where the output index is to be written.
  Defaults to "maxima-index-html.lisp", as before
* truenamep: When set to T, the html index file will include the full
  path to the file.  Defaults to NIL.  This is intended to be used
  when adding external docs for display.

To make this work, pass the `truenamep` parameter to `process-line`
which will either just use the filename or use the truename of the
file.

In `display-html-help`, if the path contains a directory, just use it
as is; otherwise, we prefix the path with `$url_base`, as before.