Package stringproc: set assign property for variables newline, tab, and space to require that an assigned value must be a string.
This is a direct translation of the previous (nonfunctional) DEFMVAR formulation
which attempted (unsuccessfully) to specify that those variables are strings.

I am not convinced that those variables are needed at all, but this is a
minimal change which maintains the previous concept.
I think we should consider the following: making those variables not assignable;
changing the names to something like string_newline, etc., to reduce
the probability of collision; or removing them entirely.