Rename tab to datatab in numeric/interpol.mac
The variable `tab` is defined in stringproc and only allows strings to
be assigned to it.  interpol.mac uses `tab` as a parameter for a
matrix and this conflicts with the restriction for `tab`.

To resolve this, rename `tab` to `datatab` as the least disruptive way
to fix this.

This also fixes bug #4379.