tests/command-line-options.el: test syntax error is fatal
Add two tests to verify the change made in commit 93cfd0dee that makes
a syntax error fatal in conjunction with --quit-on-error.

Note: GCL is currently marked as failing the tests
- syntax-error2
- string
- string2

In the first test, there is a missing space after the newline; in
the last two, there is a missing newline at the end of the first part
of the error message. These are, I understand, due to a bug in the way
that GCL handles MFORMAT strings [1].

Ref:
[1] https://lists.gnu.org/archive/html/gcl-devel/2024-03/msg00014.html