Rename --no-verify-html-index and verify only when switching to html
Rename `--no-verify-html-index` to `--verify-html-index`, which
defaults to false.  Thus, the index is no longer verified on startup.

However, when switching the output format for help, we now always verify the
index so that the user knows that something may be wrong with the html
help.

Finally, when building the docs, add `--verify-html-index` so that we
get notified if something is wrong with the index.