Fix #4414:  Add url for bug reporting page in bug_report
Also moved the entries related to the test suite at the end of the
section.  I think it makes more sense this way since the section is
about bug reporting, so the bug reporting stuff should be
first.  (This is really only relevant if you're reading the section in
the manual in its entirety.   Help should only show exactly the
desired help item, of course.)