Fix possible lisp error when translating entier
When entier's arg had a mode of float or number, the translator has
attempted to check if the arg is coming from a sqrt call and
incorrectly assumed it could take the car of the translated arg.

(%i1) foo () := entier (2.0)$

(%i2) translate (foo);
<lisp error>

Even if the translated arg was a cons, the sqrt check is outdated and
wouldn't do anything useful.  The translations for entier and friends
need work, so just remove this bogus case.

No problems with the test suite, share test suite or rtest_translator.
New tests have been added to rtest_translator.