Define new macro WITH-TOP-LEVEL-ENVIRONMENT, and use it to define MEVAL*.
WITH-TOP-LEVEL-ENVIRONMENT just does exactly the same stuff as MEVAL*
did before, so MEVAL* has just the same effect as before.

The macro is intended to make the effect of calling Maxima functions
from Lisp the same as evaluating the functions in the Maxima repl,
since the repl evaluates input expressions via MEVAL*.

In particular, WITH-TOP-LEVEL-ENVIRONMENT (previously MEVAL*)
initializes VARLIST, which is referenced by several Maxima functions.

Partial solution for SF bug #4441: "gf_factor in ASDF build"