Use def-simp-mqapply for simplifier for psi
Comment out the setting of the specsimp property for $psi and the
functions psisimp and psisimp1.

Use def-simp-mqapply to define the simplifier for psi.  This requires
moving the function psisimp1 into the body of the simplifier itself as
a labels function so that we can give up appropriately.

Full testsuite + share tests pass.