Fix #4454:  Support * in to_lisp() mode
We add support for the special variables `*` and friends in
`to_lisp()` mode so that you can refer to previous results as you
could in a normal CL repl.

This includes `*`, `**`, `***`, `/`, `//`, `///`, `+`, `++`,
`+++`. and `-`.

The function `interactive-eval` is taken from cmucl and very slightly
modified to work with `maxima-read-eval-print-loop`.

Manually tested this with cmucl and ccl64 (ccl64 results below):
```
(%i1) to_lisp();

Type (to-maxima) to restart, ($quit) to quit Maxima.

MAXIMA> (+ 4 5)

9
MAXIMA> *
9
MAXIMA> (+ 1 2)

3
MAXIMA> (+ 2 3)

5
MAXIMA> **
3
MAXIMA> (truncate 7 2)

3
1
MAXIMA> /
(3 1)
MAXIMA> **
3
MAXIMA> +
**
```