Make SIMPMIN/SIMPQUOT use SIMPLIFYA correctly
SIMPMIN and SIMPQUOT were always calling SIMPLIFYA on their arguments,
even if they were already simplified (as indicated by the third argument Z).

Now, if Z is T (arguments are already simplified), they are taken as-is.
SIMPLIFYA is only called if Z is NIL. That's how the other simplification
functions do it.